and-ask-src/requirements.txt
Django==4.2.5
djangorestframework==3.15.0
#django-environ==0.11.0
django-filter==23.1
django-autoslug==1.9.8
django-countries==7.3
Pillow==10.0.0
django-phonenumber-field==7.0.0
phonenumbers==8.13.19
psycopg2-binary==2.9.8
flake8==6.1.0
black==23.7.0
isort==5.12.0
python-dotenv==1.0.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.8.0
pytest-django==4.6.0
pytest-factoryboy==2.4.0
Faker==19.12.0
pytest-cov==4.1.0
mysqlclient
pandas==2.1.0 
numpy==1.24.0
openpyxl==3.1.2
xlrd==2.0.1


